requirement.txt
djangorestframework = 3.12.4
django = 3.2.4
djangorestframework-simplejwt = 4.7.1
django-rest-knox = 4.1.0